Conflicts:
	goobi-viewer-core/.settings/org.eclipse.wst.common.component
	goobi-viewer-core/WebContent/WEB-INF/lib/normdataimporter.jar

+++ goobi-viewer-core/src/de/intranda/digiverso/presentation/managedbeans/CmsMediaBean.java
@@ -83,7 +83,7 @@ public class CmsMediaBean implements Serializable {
                     saveMedia();
                 } catch (IOException | DAOException e) {
                     logger.error("Failed to upload media file:{}", e.getMessage());
-                    if (mediaFile.isFile()) {
+                    if (mediaFile != null && mediaFile.isFile()) {
                         mediaFile.delete();
                     }
                 }
